tidying up UI and fixing one station in the dataset

+++ src/main/java/org/neo4j/tube/Instruction.java
@@ -30,9 +30,9 @@ public class Instruction
         return station;
     }
 
-    public double getDuration()
+    public int getDuration()
     {
-        return duration;
+        return (int) duration;
     }
 
     @Override
